+++ src/mlz-pack.ts
@@ -23,19 +23,21 @@ function printBanner () {
 
 program
   .version(`mlz-pack ${pkg.version}`)
-  .usage('<command> [options]')
+  .usage('<command> [options]');
 
 program
-  .command('build [env]')
-  .description('build your project in dev/prod mode')
-  .option('-e, --entry <entry>', 'entry file(.js||.ts||.tsx), default: src/index')
+  .command('build [entry]')
+  .description('build your project from a entry file(.js||.ts||.tsx), default: src/index.tsx')
+  .option('-e, --env <environment>', 'dev or prod（default: prod）')
   .option('-d, --dest <dest>', 'output directory (default: build)')
-  .action((env, cmd) => {
-    if (env === 'dev') {
-      Init.build('dev');
-    } else {
-      Init.build('prod');
-    }
+  .action((entry, cmd) => {
+    const config = {
+      webpack: {
+        entryPath: entry,
+        buildPath: cmd.dest,
+      },
+    };
+    Init.build(cmd.env === 'dev' ? 'dev' : 'prod', config);
   });
 
 program
@@ -43,13 +45,25 @@ program
   .description('serve your project in development mode')
   .option('-p, --port <port>', 'port used by the server (default: 8080)')
   .action((entry, cmd) => {
-    Init.serve();
-  })
-
+    const config = {
+      webpack: {
+        entryPath: entry,
+        devServer: {
+          port: cmd.port,
+        },
+      },
+    };
+    Init.serve(config);
+  });
+program
+  .arguments('<command>')
+  .action((cmd) => {
+    program.outputHelp();
+    console.log(`  ` + chalk.red(`Unknown command ${chalk.yellow(cmd)}.`));
+  });
 
 program.parse(process.argv);
 
 if (!process.argv.slice(2).length) {
-  program.outputHelp()
+  program.outputHelp();
 }
-