+++ src/webpack/config.ts
@@ -1,62 +1,32 @@
-const path = require('path');
+import path from 'path';
 
-export type WebpackConfig = {
-  isDev:boolean;
-  rootPath:string;
-  entryPath:string[];
-  buildPath:string;
-  publicPath:string;
-  devServer: {
-    port:string;
-  },
-  htmlPlugin:{
-    template?:string;
-    favicon?:string;
-    filename?:string;
-    options?:{[key:string]: any};
-  },
-  analyze:boolean;
-  cssScopeName:string,
-  libs: {[key:string]: string[]};
-  alias?:{[key:string]: string};
-  definePlugin?:{[key:string]: any};
-  pxToRemPlugin?:{
-    rootValue?:number;
-    blacklist?:string[];
-  };
-};
+import { WebpackConfig } from '../types';
 
 class Config {
-  private config : WebpackConfig = {
+  private config:WebpackConfig = {
     isDev: process.env.NODE_ENV !== 'production',
+    rootPath: process.cwd(),
+    entryPath: { index: path.resolve(process.cwd(), 'src/index.tsx') },
+    buildPath: path.resolve(process.cwd(), 'build'),
+    publicPath: '/',
     devServer: {
-      port: '8080'
+      port: '8080',
     },
-    rootPath: '',
-    entryPath: [''],
+    cssScopeName: '[path][name]__[local]',
+    analyzePlugin: false,
     htmlPlugin: {
       filename: 'index.html',
-      options: {},
-    },
-    buildPath: '',
-    publicPath: '/',
-    analyze: false,
-    cssScopeName: '[name]__[hash:5]',
-    libs: {
-      vender: [],
+      template: path.resolve(process.cwd(), 'src/index.ejs'),
     },
   };
 
   public init(param?:Partial<WebpackConfig>) {
-    this.config.rootPath = process.cwd();
-    this.config.entryPath = path.join(this.config.rootPath, 'src/index');
-    this.config.buildPath = path.join(this.config.rootPath, 'build');
-    this.config.cssScopeName = this.config.isDev ? '[name]__[local]' : '[name]__[hash:5]';
+    this.config.cssScopeName = this.config.isDev ? this.config.cssScopeName : '[name]__[hash:base64:5]';
     if (param) {
       Object.assign(this.config, param);
     }
   }
-  
+
   public get() : WebpackConfig {
     return this.config;
   }