+++ src/webpack/dev.config.ts
@@ -1,31 +1,98 @@
-const webpack = require('webpack');
-const path = require('path');
-const merge = require('webpack-merge');
+import webpack from 'webpack';
+import path from 'path';
+import autoprefixer from 'autoprefixer';
+import pxtorem from 'postcss-pxtorem';
+import { CleanWebpackPlugin } from 'clean-webpack-plugin';
+import ProgressBarPlugin from 'progress-bar-webpack-plugin';
+import FriendlyErrorsWebpackPlugin from 'friendly-errors-webpack-plugin';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const HtmlWebpackTagsPlugin = require('html-webpack-tags-plugin');
 
 import { config as configs } from './config';
+import { getBabelConfig } from './babel';
 
 export const devCfg = () => {
   const config = configs.get();
   return {
-    entry: {
-    },
+    entry: config.entryPath,
     output: {
+      // 打包输出的文件
+      path: config.buildPath,
+      publicPath: config.publicPath,
       filename: 'js/[name].js',
       chunkFilename: 'js/[name].js',
-      pathinfo: false,
+    },
+    resolve: {
+      modules: [
+        config.rootPath,
+        'node_modules',
+      ],
+      alias: {
+        'root': config.rootPath,
+        ...config.alias,
+      },
+      extensions: ['.ts', '.tsx', '.js', '.css', '.scss'],
+      symlinks: false,
+      cacheWithContext: false,
     },
     mode: 'development',
     devtool: 'cheap-module-eval-source-map',
     module: {
       rules: [
         {
+          test: /\.css$/,
+          include: /node_modules/,
+          use: [
+            'style-loader',
+            'css-loader',
+            {
+              loader: 'postcss-loader',
+              options: {
+                plugins: () => {
+                  return [
+                    autoprefixer(),
+                  ];
+                },
+              },
+            },
+            'sass-loader',
+          ],
+        },
+        {
+          test: /\.s?css$/,
+          exclude: /node_modules/,
+          use: [
+            { loader: 'style-loader' },
+            {
+              loader: 'css-loader',
+              options: {
+                modules: {
+                  localIdentName: config.cssScopeName,
+                },
+              },
+            },
+            {
+              loader: 'postcss-loader',
+              options: {
+                plugins: () => {
+                  const plugin = [autoprefixer()];
+                  if (config.pxtorem) {
+                    plugin.push(pxtorem(config.pxtorem));
+                  }
+                  return plugin;
+                },
+              },
+            },
+            'sass-loader',
+          ],
+        },
+        {
           test: /\.(woff|woff2|eot|ttf|mp3)$/,
           exclude: /(node_modules)/,
           use: [
             {
-              loader: require.resolve('file-loader'),
+              loader: 'file-loader',
             },
           ],
         },
@@ -34,7 +101,7 @@ export const devCfg = () => {
           exclude: /node_modules/,
           use: [
             {
-              loader: require.resolve('url-loader'),
+              loader: 'url-loader',
               options: {
                 limit: 100000,
                 name: 'img/[hash].[ext]',
@@ -42,14 +109,44 @@ export const devCfg = () => {
             },
           ],
         },
+        {
+          test: /\.(ts|tsx)?$/,
+          use: [
+            {
+              loader: 'babel-loader',
+              options: getBabelConfig(),
+            },
+          ],
+          exclude: /(node_modules)/,
+        },
+        {
+          test: /\.worker\.js$/,
+          use: {
+            loader: 'worker-loader',
+            options: {
+              name: '[name].js',
+              inline: true,
+            },
+          },
+          exclude: /(node_modules)/,
+        },
       ],
     },
     plugins: [
+      new ProgressBarPlugin(),
+      new FriendlyErrorsWebpackPlugin(),
+      new CleanWebpackPlugin({
+        verbose: true, // Write logs to console.
+        dry: false, // Use boolean 'true' to test/emulate delete. (will not remove files).
+      }),
       new webpack.DefinePlugin({
         'DEBUG': true,
         ...config.definePlugin,
       }),
-      new webpack.HotModuleReplacementPlugin(),
+      new HtmlWebpackPlugin({
+        ...config.htmlPlugin,
+      }),
+      new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
       // TODO 的环境dll加速build
       // new webpack.DllReferencePlugin({
       //   context: path.join(__dirname, 'dll'),