+++ src/webpack/product.config.ts
@@ -1,19 +1,40 @@
-const webpack = require('webpack');
+import webpack from 'webpack';
 import TerserPlugin from 'terser-webpack-plugin';
-const CompressionPlugin = require("compression-webpack-plugin")
-const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
+import CompressionPlugin from 'compression-webpack-plugin';
+import OptimizeCSSAssetsPlugin from 'optimize-css-assets-webpack-plugin';
+import MiniCssExtractPlugin from 'mini-css-extract-plugin';
+import { CleanWebpackPlugin } from 'clean-webpack-plugin';
+import autoprefixer from 'autoprefixer';
+import pxtorem from 'postcss-pxtorem';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
 
 import { config as configs } from './config';
+import { getBabelConfig } from './babel';
 
-export const prodCfg = () =>{
+export const prodCfg = () => {
   const config = configs.get();
   const libraries = config.libs;
   return {
     mode: 'production',
+    entry: config.entryPath,
     output: {
+      path: config.buildPath,
       publicPath: config.publicPath, // local: '/'
       filename: 'js/[name].[chunkhash].js',
-      chunkFilename: 'js/[name].[chunkhash].js'
+      chunkFilename: 'js/[name].[chunkhash].js',
+    },
+    resolve: {
+      modules: [
+        config.rootPath,
+        'node_modules',
+      ],
+      alias: {
+        'root': config.rootPath,
+        ...config.alias,
+      },
+      extensions: ['.ts', '.tsx', '.js', '.css', '.scss'],
+      symlinks: false,
+      cacheWithContext: false,
     },
     optimization: {
       removeAvailableModules: true,
@@ -23,54 +44,113 @@ export const prodCfg = () =>{
         name: 'manifest',
       },
       splitChunks: {
-      chunks: 'all',
-      cacheGroups: {
-          vendor: {
+        chunks: 'all',
+        cacheGroups: {
+          vendors: {
             test: /node_modules/,
             chunks: 'all',
             name(module) {
-              const packageName = module.context.match(/[\\/]node_modules[\\/](.*?)([\\/]|$)/)[1];
-              const names = Object.keys(libraries);
-              let name = 'lib';
-              names.map((val) => {
-                if (libraries[val].indexOf(packageName) >= 0) {
-                  name = val;
-                }
-              });
+              let name = 'venderLibs';
+              if (libraries) {
+                const packageName = module.context.match(/[\\/]node_modules[\\/](.*?)([\\/]|$)/)[1];
+                const names = Object.keys(libraries);
+                names.map((val) => {
+                  if (libraries[val].indexOf(packageName) >= 0) {
+                    name = val;
+                  }
+                });
+              }
               return name;
-            }
-          }
-        }
+            },
+          },
+        },
       },
       minimizer: [
-        new TerserPlugin({
-          sourceMap: true,
-        }),
+        new TerserPlugin({ sourceMap: true }),
         new OptimizeCSSAssetsPlugin({
           cssProcessorOptions: {
             discardComments: { removeAll: true },
           },
-          canPrint: true
+          canPrint: true,
         }),
-      ]
+      ],
     },
     module: {
       rules: [
+        {
+          test: /\.css$/,
+          include: /node_modules/,
+          use: [
+            'style-loader',
+            'css-loader',
+            {
+              loader: 'postcss-loader',
+              options: {
+                plugins: () => {
+                  return [
+                    autoprefixer(),
+                  ];
+                },
+              },
+            },
+            'sass-loader',
+          ],
+        },
+        {
+          test: /\.s?css$/,
+          exclude: /node_modules/,
+          use: [
+            {
+              loader: MiniCssExtractPlugin.loader,
+            },
+            {
+              loader: 'css-loader',
+              options: {
+                modules: {
+                  localIdentName: config.cssScopeName,
+                },
+              },
+            },
+            {
+              loader: 'postcss-loader',
+              options: {
+                plugins: () => {
+                  const plugin = [autoprefixer()];
+                  if (config.pxtorem) {
+                    plugin.push(pxtorem(config.pxtorem));
+                  }
+                  return plugin;
+                },
+              },
+            },
+            'sass-loader',
+          ],
+        },
+        {
+          test: /\.(ts|tsx)?$/,
+          use: [
+            {
+              loader: 'babel-loader',
+              options: getBabelConfig(),
+            },
+          ],
+          exclude: /(node_modules)/,
+        },
         {
           test: /\.(jpe?g|png|gif|svg)$/,
           exclude: /node_modules/,
           use: [
             {
-              loader: require.resolve('url-loader'),
+              loader: 'url-loader',
               options: {
                 limit: 3 * 1024,
                 name: 'images/[name]_[hash:5].[ext]',
                 publicPath: config.publicPath,
-              }
+              },
             },
-            // {
-            //   loader: require.resolve('image-webpack-loader'), //图片压缩
-            // }
+            {
+              loader: require.resolve('image-webpack-loader'), // 图片压缩
+            },
           ],
         },
         {
@@ -78,25 +158,45 @@ export const prodCfg = () =>{
           exclude: /node_modules/,
           use: [
             {
-              loader: require.resolve('file-loader'),
+              loader: 'file-loader',
               options: {
                 name: 'assets/[name]_[hash:5].[ext]',
                 publicPath: config.publicPath,
-              }
-            }
+              },
+            },
           ],
-        }
-      ]
+        },
+        {
+          test: /\.worker\.js$/,
+          use: {
+            loader: 'worker-loader',
+            options: {
+              name: '[name].js',
+              inline: true,
+            },
+          },
+          exclude: /(node_modules)/,
+        },
+      ],
     },
     plugins: [
+      new CleanWebpackPlugin({
+        verbose: true, // Write logs to console.
+        dry: false, // Use boolean 'true' to test/emulate delete. (will not remove files).
+      }),
       new webpack.DefinePlugin({
         'DEBUG': false,
         ...config.definePlugin,
       }),
+      new MiniCssExtractPlugin({
+        filename: '[name].[contenthash].css',
+        chunkFilename: '[id].[contenthash].css',
+        ignoreOrder: false, // Enable to remove warnings about conflicting order
+      }),
       new CompressionPlugin({
         test: /\.js$|\.css$|\.html$/,
         threshold: 1024,
-        minRatio: 0.8
+        minRatio: 0.8,
       }),
       new webpack.SourceMapDevToolPlugin({
         // TODO sourceMap的地址
@@ -104,6 +204,16 @@ export const prodCfg = () =>{
         // publicPath: config.SOURCEMAP,
         filename: '[file].map',
       }),
-    ]
+      new HtmlWebpackPlugin({
+        ...config.htmlPlugin,
+        removeAttributeQuotes: true,
+        collapseWhitespace: true,
+        html5: true,
+        minifyCSS: true,
+        removeComments: false,
+        removeEmptyAttributes: true,
+      }),
+      new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+    ],
   };
-}
+};