+++ test/smoke/template-project/mlz-pack.js
@@ -1,17 +1,21 @@
 const path = require('path');
 
 module.exports = {
-  "baseUrl": __dirname,
-  "entry": [path.join(__dirname, './src/index.ts')],
-  html:{
-    filename: 'index.html',
-  },
-  alias: {
-    src: path.join(__dirname, 'src'),
-  },
-  globalVariable: {
-    Debug: false
-  },
-  analyze: false,
-  port: 5000,
+  webpack: {
+    "rootPath": __dirname,
+    "entryPath": {index: path.join(__dirname, './src/index.ts')},
+    devServer: {
+      port: 4000,
+    },
+    alias: {
+      src: path.join(__dirname, 'src'),
+    },
+    definePlugin: {
+      Debug: false
+    },
+    analyzePlugin: false,
+    htmlPlugin: {
+      template: path.join(__dirname, 'src/index.html'),
+    }
+  }
 }