+++ CHANGELOG.md
@@ -2,6 +2,105 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.15](https://github.com/juicecube/mlz-pack/compare/v0.0.14...v0.0.15) (2019-10-31)
+
+
+### Bug Fixes
+
+* make serve use dev config ([ef90a8c](https://github.com/juicecube/mlz-pack/commit/ef90a8c))
+
+### [0.0.14](https://github.com/juicecube/mlz-pack/compare/v0.0.13...v0.0.14) (2019-10-31)
+
+
+### Bug Fixes
+
+* use relatetive path in cli entry ([f5784ab](https://github.com/juicecube/mlz-pack/commit/f5784ab))
+* use relatetive path in cli entry ([09c7f97](https://github.com/juicecube/mlz-pack/commit/09c7f97))
+
+### [0.0.13](https://github.com/juicecube/mlz-pack/compare/v0.0.12...v0.0.13) (2019-10-31)
+
+
+### Bug Fixes
+
+* css-laoder in production mode ([fe88237](https://github.com/juicecube/mlz-pack/commit/fe88237))
+* css-loader in production ([4c8d7a1](https://github.com/juicecube/mlz-pack/commit/4c8d7a1))
+
+### [0.0.12](https://github.com/juicecube/mlz-pack/compare/v0.0.11...v0.0.12) (2019-10-31)
+
+
+### Bug Fixes
+
+* remove console.log and add entry config in serve ([8bb6719](https://github.com/juicecube/mlz-pack/commit/8bb6719))
+
+### [0.0.11](https://github.com/juicecube/mlz-pack/compare/v0.0.10...v0.0.11) (2019-10-30)
+
+
+### Bug Fixes
+
+* css-loader ([b364687](https://github.com/juicecube/mlz-pack/commit/b364687))
+
+### [0.0.10](https://github.com/juicecube/mlz-pack/compare/v0.0.9...v0.0.10) (2019-10-30)
+
+
+### Bug Fixes
+
+* antd and local moment ([20e3b44](https://github.com/juicecube/mlz-pack/commit/20e3b44))
+
+### [0.0.9](https://github.com/juicecube/mlz-pack/compare/v0.0.8...v0.0.9) (2019-10-30)
+
+
+### Bug Fixes
+
+* css loader ([c05d558](https://github.com/juicecube/mlz-pack/commit/c05d558))
+
+### [0.0.8](https://github.com/juicecube/mlz-pack/compare/v0.0.7...v0.0.8) (2019-10-30)
+
+
+### Bug Fixes
+
+* remove hot plugin and add image-loader ([31c50dc](https://github.com/juicecube/mlz-pack/commit/31c50dc))
+
+### [0.0.7](https://github.com/juicecube/mlz-pack/compare/v0.0.6...v0.0.7) (2019-10-30)
+
+
+### Bug Fixes
+
+* remove require.resolve and change readme ([1e92bab](https://github.com/juicecube/mlz-pack/commit/1e92bab))
+
+### [0.0.6](https://github.com/juicecube/mlz-pack/compare/v0.0.5...v0.0.6) (2019-10-29)
+
+
+### Bug Fixes
+
+* load option config ([92e8e56](https://github.com/juicecube/mlz-pack/commit/92e8e56))
+
+### [0.0.5](https://github.com/juicecube/mlz-pack/compare/v0.0.4...v0.0.5) (2019-10-29)
+
+
+### Bug Fixes
+
+* update webpack config ([0715d68](https://github.com/juicecube/mlz-pack/commit/0715d68))
+
+### [0.0.4](https://github.com/juicecube/mlz-pack/compare/v0.0.3...v0.0.4) (2019-10-28)
+
+
+### Bug Fixes
+
+* add .npmignore ([84a24e6](https://github.com/juicecube/mlz-pack/commit/84a24e6))
+
+### [0.0.3](https://github.com/juicecube/mlz-pack/compare/v0.0.2...v0.0.3) (2019-10-28)
+
+
+### Bug Fixes
+
+* css-loader bug ([55d3d77](https://github.com/juicecube/mlz-pack/commit/55d3d77))
+* resolve conflict ([df4f1f5](https://github.com/juicecube/mlz-pack/commit/df4f1f5))
+
+
+### Features
+
+* fix tslint code ([c51f3bf](https://github.com/juicecube/mlz-pack/commit/c51f3bf))
+
 ### 0.0.2 (2019-10-23)